templates/crewai-agent-template/requirements.txt
crewai>=0.50.0
nerq>=1.0.0
python-dotenv>=1.0.0
